redis-client/requirements.txt
python-dotenv~=1.0.1
aiohttp~=3.9.5
pydantic~=2.8.2
pydantic-settings~=2.3.4
APScheduler~=3.10.4
redis==5.0.1